Move LoveObject and Conf to types/
