+++ include/types/Conf.d.ts
